move payout tests to separate file
